Add -channels flag to split the messages across channels.
It’s not very typical for an IRC channel to receive thousands of
messages per second, so this flag makes the benchmark a bit more
realistic.

The most important difference is that the number of messages which the
server needs to hand out is reduced — before, we had *numSessions
sessions in the same channel, e.g. 250. For every message, the server
had to send 249 messages to all the other sessions. Now, it will need to
send 49 messages by default, due to *numChannels defaulting to
*numSessions / 50.