Rewrite throughput.go, move to cmd/throughput
The new version exports prometheus metrics instead of writing GNUplot
files.

Further, instead of connecting to a bridge, it directly initiates
robustsessions (after verifying the network is healthy).

throughput sends as many messages as possible and exits once the spread
between the minimum/maximum messages/s over the last 10s converges to
less than 10% of the maximum messages/s.

When done, the specified grafana dashboards will be snapshotted to
snapshot.raintank.io

grafana/ lives in internal/ because we don’t want third-party packages
to make use of this code. In case you want to use this code, please fork
it and maintain it :).

related to https://github.com/robustirc/robustirc/issues/81